Ver 2.0.1.05  Query fixes for Version 2 Files modified - [Job1111,.wmproject.properties,wmProperties.js]

+++ services/Job1111/src/com/psabackoffice/job1111/models/query/GetBidActivityQuantitesEquipResponse.java
@@ -19,7 +19,7 @@ public class GetBidActivityQuantitesEquipResponse implements Serializable {
     private Integer bidId;
 
     @ColumnAlias("TakeOffQuantity")
-    private BigDecimal takeOffQuantity;
+    private Double takeOffQuantity;
 
     @ColumnAlias("QuantityUnit")
     private String quantityUnit;
@@ -38,11 +38,11 @@ public class GetBidActivityQuantitesEquipResponse implements Serializable {
         this.bidId = bidId;
     }
 
-    public BigDecimal getTakeOffQuantity() {
+    public Double getTakeOffQuantity() {
         return this.takeOffQuantity;
     }
 
-    public void setTakeOffQuantity(BigDecimal takeOffQuantity) {
+    public void setTakeOffQuantity(Double takeOffQuantity) {
         this.takeOffQuantity = takeOffQuantity;
     }
